+++ src/utils/index.ts
@@ -1,13 +1,13 @@
 export {
   convertUtcToLocalTime,
-  convertUtcToFullLocalTime,
+  convertUtcToFullLocalTime
 } from "./dateTimeUtils";
 
 export {
   setClipBoardText,
   isObjectHaveSameData,
   generateGUID,
-  openWindow,
+  openWindow
 } from "./funcUtils";
 
 export { validateTokenExpireTime } from "./jwtUtils";
@@ -15,5 +15,3 @@ export { validateTokenExpireTime } from "./jwtUtils";
 export { getQueryString } from "./routerUtils";
 
 export { getDocumentTitle } from "./styleUtils";
-
-export { default as history } from "./history";